refactor: UserController 및 UserService의 로직을 AuthController/AuthService로 이동
