+++ domain/src/main/java/com/pickcar/auth/presentation/AuthController.java
@@ -1,9 +1,12 @@
 package com.pickcar.auth.presentation;
 
 import com.pickcar.auth.application.AuthService;
+import com.pickcar.auth.application.UserService;
 import com.pickcar.auth.presentation.dto.request.AuthRequest;
+import com.pickcar.auth.presentation.dto.request.UserInfoRequest;
 import com.pickcar.auth.presentation.dto.response.AccessTokenResponse;
 import com.pickcar.auth.presentation.dto.response.AuthResponse;
+import com.pickcar.auth.presentation.dto.response.EmployeeListResponse;
 import com.pickcar.security.jwt.JwtConstants;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
@@ -13,6 +16,8 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.List;
+
 @Slf4j
 @RestController
 @RequiredArgsConstructor
@@ -20,6 +25,20 @@ import org.springframework.web.bind.annotation.*;
 public class AuthController {
 
     private final AuthService authService;
+    private final UserService userService;
+
+    @PostMapping("/sign-up")
+    @ResponseStatus(HttpStatus.CREATED)
+    public void register(@RequestBody UserInfoRequest request) {
+        authService.create(request);
+    }
+
+    @GetMapping("/employees")
+    @ResponseStatus(HttpStatus.OK)
+    public List<EmployeeListResponse> getEmployees() {
+        List<EmployeeListResponse> responses = authService.getAllEmployees();
+        return responses;
+    }
 
     @PostMapping("/login")
     @ResponseStatus(HttpStatus.OK)
@@ -30,8 +49,6 @@ public class AuthController {
         setAccessTokenCookie(response, authResponse.accessToken());
         return new AccessTokenResponse(authResponse.accessToken());
     } //TODO: 로그인 실패 예외처리 추가
-
-
 
     @PostMapping("/logout")
     @ResponseStatus(HttpStatus.OK)
